sqlanywhere-mcp/pyproject.toml
[project]
name = "sqlanywhere-mcp"
version = "0.1.0"
description = "MCP server for SQL Anywhere database connectivity via ODBC"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=0.9.0",
    "pyodbc>=5.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
sqlanywhere-mcp = "sqlanywhere_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["sqlanywhere_mcp"]
